+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.85.1
 
+- Backups / SVRB - add support for multiple SVRB backends when new enclaves need to roll out.